saving packets and compts code reorganization #2,  not working yet

+++ src/main/java/com/somecode/dao/ComptDaoImpl.java
@@ -87,8 +87,7 @@ public class ComptDaoImpl implements  ComptDao {
     public List<State> getAllStates() {
         Iterable<State> iterable = stateRepository.findAll();
         states = Lists.newArrayList(iterable);
-        System.out.println("states " + states);
-
+        LOGGER.info(" All states " + states);
         return states;
     }
 
@@ -109,6 +108,7 @@ public class ComptDaoImpl implements  ComptDao {
         return allComboData;
     }
 
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
     private boolean checkComboDataListsForEquality(List<ComboData> allComboData,
                                                    List<ComboData> oldAllComboData) {
         int size = allComboData.size();
@@ -120,6 +120,7 @@ public class ComptDaoImpl implements  ComptDao {
                         getSortedLabels(allComboData).get(i).equals(getSortedLabels(oldAllComboData).get(i)));
     }
 
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
     private List<String> getSortedLabels(List<ComboData> comboDataList) {
         return comboDataList.stream().map(ComboData::getLabel).sorted().collect(Collectors.toList());
     }
@@ -139,22 +140,22 @@ public class ComptDaoImpl implements  ComptDao {
         return getCompts;
     }
 
-    @Transactional(propagation = Propagation.NOT_SUPPORTED)
+    @Transactional
     private List<Integer> getIndicesFromVals(List<String> vals) {
         getAllComboData();
         return vals.stream().map(mapComboLabelsToIndices::get).collect(Collectors.toList());
     }
 
-    @Override
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public List<Long> updateCompts(List<ComptParams> comptParamsList) {
+    @Transactional
+    private List<Long> updateCompts(List<ComptParams> comptParamsList, Long packetId) {
         List<Long> result = new ArrayList<>();
 
         for (ComptParams comptParams : comptParamsList) {
             long comptId = comptParams.getId();
             Compt compt = em.find(Compt.class, comptId, LockModeType.OPTIMISTIC_FORCE_INCREMENT);
             if (compt == null) {
-                LOGGER.error("Compt update. The compt with id " + comptId + "does not exist.");
+                LOGGER.error("Packets updating or saving. Compt update. " +
+                        "The compt with id " + comptId + "does not exist.");
                 continue;
             }
             List<Integer> newCheckedIndices = getIndicesFromVals(comptParams.getVals());
@@ -166,36 +167,13 @@ public class ComptDaoImpl implements  ComptDao {
                 if (!checked && newCheckedIndices.get(stateIndex) == comboDataIndex
                         || checked && newCheckedIndices.get(stateIndex) != comboDataIndex) {
                     dc.setChecked(!checked);
-                    LOGGER.info("DataCompt updated widh id = " + dc.getId());
+                    LOGGER.info("Packets updating or saving. Compt update. " +
+                            "Compt#" + comptId + ": DataCompt updated widh id = " + dc.getId());
                 }
             }
             result.add(compt.getId());
         }
-        LOGGER.info("List of the updated compts: " + result);
-        return result;
-    }
-
-    @Override
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public List<Long> updatePackets(List<PacketParams> updatedPackets) {
-        List<Long> result = new ArrayList<>();
-        for (PacketParams p : updatedPackets) {
-            long packetId = p.getId();
-            Packet packet = em.find(Packet.class, packetId, LockModeType.OPTIMISTIC_FORCE_INCREMENT);
-            if (packet == null) {
-                LOGGER.error("Packet state update. The packet with id " + packetId + "does not exist.");
-                continue;
-            }
-            long newStateId = p.getStateId();
-            State newState = em.find(State.class, newStateId);
-            if (newState == null) {
-                LOGGER.error("Packet state update. The state with id " + newStateId + "does not exist.");
-                continue;
-            }
-            packet.setState(newState);
-            result.add(packetId);
-        }
-        LOGGER.info("List of the updated packet ids: " + result);
+        LOGGER.info("Packet#" + packetId + ": List of the updated compts: " + result);
         return result;
     }
 
@@ -208,12 +186,27 @@ public class ComptDaoImpl implements  ComptDao {
                 .filter(c -> c != null)
                 .collect(Collectors.toList());
 
+        int comptsSize = compts.size();
+        if (comptsSize == 0) {
+            LOGGER.error("Compts' delete. Can't delete any of the compts with the following(ALL GIVEN TO METHOD)" +
+                    " ids cause they don't exist: " + idsToDelete);
+            return Collections.EMPTY_LIST;
+        }
+
         comptRepository.delete(compts);
         List<Long> result = getIdsFromEntities(compts.toArray(new Compt[0]));
-        LOGGER.info("Deleted components with the following ids: " + result);
+        LOGGER.info("Deleted compts with the following ids: " + result);
+
+        if (comptsSize != idsToDelete.size()) {
+            Set<Long> idsToDeleteSet = new HashSet<>(idsToDelete);
+            LOGGER.error("Packets' delete. Can't delete any of the packets with the following ids cause " +
+                    "they don't exist: " + idsToDeleteSet.removeAll(result));
+        }
+
         return result;
     }
 
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
     private List<Long> getIdsFromEntities(EntityType[] entities) {
         return Arrays.stream(entities).mapToLong(EntityType::getId).boxed().collect(Collectors.toList());
     }
@@ -221,55 +214,90 @@ public class ComptDaoImpl implements  ComptDao {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public List<Long> deletePackets(List<Long> idsToDelete) {
+
         List<Packet> packets = packetRepository
                 .findByIdIn(idsToDelete)
                 .stream()
                 .filter(p -> p != null)
                 .collect(Collectors.toList());
-        if (packets.size() == 0) {
-            LOGGER.error("Can't delete any of the packets with the following ids cause " +
-                    "they don't exist: " + idsToDelete);
+
+        int packetsSize = packets.size();
+        if (packetsSize == 0) {
+            LOGGER.error("Packets' delete. Can't delete any of the packets with the following(ALL GIVEN TO METHOD)" +
+                    " ids cause they don't exist: " + idsToDelete);
             return Collections.EMPTY_LIST;
         }
         packetRepository.delete(packets);
         List<Long> result = getIdsFromEntities(packets.toArray(new Packet[0]));
-        LOGGER.info("Deleted packets with the following ids: " + result);
+        LOGGER.info("Packets' delete. Deleted packets with the following ids: " + result);
+
+        if (packetsSize != idsToDelete.size()) {
+            Set<Long> idsToDeleteSet = new HashSet<>(idsToDelete);
+            LOGGER.error("Packets' delete. Can't delete any of the packets with the following ids cause " +
+                    "they don't exist: " + idsToDeleteSet.removeAll(result));
+        }
+
         return result;
     }
 
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public List<Long> addCompts(long packetId, List<ComptParams> comptParamsList) {
-        Packet packet = getPacket(packetId);
-        if (packet == null) {
-            LOGGER.info("Packet state add compts. The packet with id " + packetId + "does not exist.");
-            return Collections.EMPTY_LIST;
+    public void saveOrUpdatePackets(List<PacketParams> createPacketParamsList,
+                                    List<PacketParams> updatePacketParamsList) {
+
+        List<Packet> packets = new ArrayList<>(createPacketParamsList.size() + updatePacketParamsList.size());
+        for (PacketParams packetParams : createPacketParamsList) {
+            long stateId = packetParams.getStateId();
+            State state = em.find(State.class, stateId);
+            if (state == null) {
+                LOGGER.info("Packets updating or saving. Adding new packets." +
+                        " The state with id " + stateId + "does not exist.");
+                continue;
+            }
+            Packet packet = new Packet(0L, state);
+            preparePacketAndComptsForSaving(packet, packetParams.getAddedComptParamsList());
+            packets.add(packet);
         }
 
-        List<Long> comptIdsList = new ArrayList<>(comptParamsList.size());
-        List<Compt> comptList = preparePacketAndComptsForSaving(packet, comptParamsList);
-        comptRepository.save(comptList).forEach(compt -> comptIdsList.add(compt.getId()));
-        LOGGER.info("Persisted new compt with ids: " + comptIdsList + " for packet#" + packetId);
-        return comptIdsList;
-    }
+        Set<Long> updatedPacketIds = new HashSet<>();
+        for (PacketParams packetParams : updatePacketParamsList) {
+            long packetId = packetParams.getId();
+            updatedPacketIds.add(packetId);
+            Packet packet = getPacket(packetId);
+            if (packet == null) {
+                LOGGER.info("Packets updating or saving. Updating the packet. " +
+                        "The packet with id " + packetId + " does not exist.");
+                continue;
+            }
+            List<ComptParams> addedComptParamsList = packetParams.getAddedComptParamsList();
+            if (addedComptParamsList.size() > 0) {
+                LOGGER.info("Packets updating or saving. The following compts for the packet#" + packetId +
+                        " were added: " + preparePacketAndComptsForSaving(packet, addedComptParamsList));
+            }
 
-    @Override
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public List<Long> addPackets(List<PacketParams> packetParamsList) {
-        List<Long> result = new ArrayList<>();
-        List<Packet> packets = new ArrayList<>(packetParamsList.size());
-        for (PacketParams packetParams : packetParamsList) {
-            State state = em.find(State.class, packetParams.getStateId());
-            Packet packet = new Packet(packetParams.getId(), state);
-            preparePacketAndComptsForSaving(packet, packetParams.getComptParamsList());
+            Long stateId = packetParams.getStateId();
+            if (stateId != null) {
+                State state = em.find(State.class, stateId);
+                if (state != null) {
+                    packet.setState(state);
+                    LOGGER.info("Packets updating or saving. The state for the packet#" + packetId +
+                            " was updated to the new one with id: " + stateId);
+                } else {
+                    LOGGER.info("Packets updating or saving. Packet state update. " +
+                            "The state with id " + stateId + "does not exist.");
+                }
+            }
+
+            updateCompts(packetParams.getUpdatedComptParamsList(), packetId);
             packets.add(packet);
         }
         packetRepository.save(packets).forEach(pkt -> {
-            LOGGER.info("Persisted new packet with id: " + pkt.getId() + " and compts with ids: " + pkt.getComptIds());
+            LOGGER.info("Packets updating or saving. Persisted new packet#" + pkt.getId() +
+                    " with the following state id: " + pkt.getState().getId() + " and compts: " + pkt.getCompts());
         });
-        return result;
     }
 
+    @Transactional
     private List<Compt> preparePacketAndComptsForSaving(Packet packet, List<ComptParams> comptParamsList) {
         List<Compt> result = new ArrayList<>(comptParamsList.size());
         List<State> statesList = getAllStates();
@@ -296,5 +324,4 @@ public class ComptDaoImpl implements  ComptDao {
         }
         return result;
     }
-
 }
